docs: clarify readme Puts the section about having Kubernetes and kubectl installed before the ark install command, which will fail if these requirements aren't met.
Removes link to the quickstart page which doesn't contain the information that it says it does,
and which is mostly redundant with this readme. There doesn't seem to be an obvious page that does
contain this information and it seems best to just use the link to the guide in the previous line.