docs: clarify readme (#239)
A small tweak to the readme to reduce confusion when installing from a
new laptop. Puts the section about having Kubernetes and kubectl
installed before the ark install command, which will fail if these
requirements aren't met.

Removes link to the quickstart page which doesn't contain the
information that the readme says it does, and which is mostly redundant
with this readme. There doesn't seem to be an obvious page that does
contain this information and it seems best to just use the link to the
guide in the previous line.

## Checklist

- [ ] Follow the [contributor guide](./CONTRIBUTING.md)
- [ ] End-to-end tests pass
- [ ] Unit tests for essential parts of your code, e.g. APIs exposed via
SDKs or endpoints
- [ ] Update `./docs`
- [ ] Recognize contributors with "@all-contributors please add <person>
for <code, bug, docs, etc>"
- [ ] Linked issues #eg101