Allow overriding default export types in config
For when there is no -t specified.
It's still necessary to have the order in which to export specified in the file, because if the resources are exported out of order they may be able to be imported. I.e. a team cannot be imported before the agents that it depends on exist.